Move `reporter` tests to a new file
